Disease Ontology (DO) Annotations

A structured annotation table mapping Disease Ontology (DOID) identifiers to human-readable disease names and definitions. This dataset is used in the pKOI framework to annotate network nodes associated with disease terms, enabling interpretation of enriched or prioritized disease-related features in biological networks.

Usage

disease_annotation

Format

A data frame with 10002 rows and 3 columns:

identifier

A character string representing the Disease Ontology ID (e.g., "DOID:0040003").

name

The name of the disease (e.g., "benzylpenicillin allergy").

definition

A detailed textual definition of the disease, including etiology, affected systems, or allergy triggers, often with references to biomedical sources.

Source

Disease Ontology https://disease-ontology.org/

Details

Disease Ontology (DO) provides a standardized, ontology-based representation of human diseases. This dataset allows integration of disease concepts into biological graphs and supports downstream analysis of disease associations.

The `identifier` column serves as a join key to nodes of type "Disease" in pKOI networks, while the `definition` column provides biological and clinical context.

Examples

data(disease_annotation)
subset(disease_annotation, grepl("allergy", name)) |> head()
#>     identifier                     name
#> 4 DOID:0040001           shrimp allergy
#> 5 DOID:0040002          aspirin allergy
#> 6 DOID:0040003 benzylpenicillin allergy
#> 7 DOID:0040004      amoxicillin allergy
#> 8 DOID:0040005      ceftriaxone allergy
#> 9 DOID:0040006    carbamazepine allergy
#>                                                                                                                                         definition
#> 4            ""A crustacean allergy that has_allergic_trigger shrimp."" [url:https\\://www.ncbi.nlm.nih.gov/pubmed/20471069] {comment=""IEDB:RV""}
#> 5     ""A drug allergy that has_allergic_trigger acetylsalicylic acid."" [url:https\\://www.ncbi.nlm.nih.gov/pubmed/2468301] {comment=""IEDB:RV""}
#> 6 ""A beta-lactam allergy that has_allergic_trigger benzylpenicillin."" [url:https\\://www.ncbi.nlm.nih.gov/pubmed/14483916] {comment=""IEDB:RV""}
#> 7      ""A beta-lactam allergy that has_allergic_trigger amoxicillin."" [url:https\\://www.ncbi.nlm.nih.gov/pubmed/11746950] {comment=""IEDB:RV""}
#> 8    ""A cephalosporin allergy that has_allergic_trigger ceftriaxone."" [url:https\\://www.ncbi.nlm.nih.gov/pubmed/12833570] {comment=""IEDB:RV""}
#> 9            ""A drug allergy that has_allergic_trigger carbamazepine."" [url:https\\://www.ncbi.nlm.nih.gov/pubmed/7602118] {comment=""IEDB:RV""}